requirements.txt
django==1.11.4
django-modern-rpc==0.8.0
djangorestframework==3.6.3
djangorestframework-jsonp==1.0.2
mysqlclient==1.3.10
pytz==2017.2
six==1.10.0
